Fixes multiple issues detected by the clang static analyzer.
I agreed with the CSA on all but one issue. I couldn't see (or fix) an
issue in esl_msa.c where the CSA thinks that useme[opos] can be used
uninitialized. I left this unsuppressed in case the reason becomes
clear later, while adding a comment in the code so I remember I looked
hard at it already.

To use the clang static analyzer:  make CC=clang CFLAGS="--analyze -g -Wall"

+++ esl_msaweight.c
@@ -1006,7 +1006,7 @@ esl_msaweight_IDFilter_adv(const ESL_MSAWEIGHT_CFG *cfg, const ESL_MSA *msa, dou
 
   /* Allocations that we always need*/
   ESL_ALLOC(sortwgt,   sizeof(double) * msa->nseq);
-  ESL_ALLOC(ranked_at, sizeof(double) * msa->nseq);
+  ESL_ALLOC(ranked_at, sizeof(int)    * msa->nseq);
   ESL_ALLOC(list,      sizeof(int)    * msa->nseq);
   ESL_ALLOC(useme,     sizeof(int)    * msa->nseq);
   esl_vec_ISet(useme, msa->nseq, 0); /* initialize array */
